+++ Sources/RichTextKit/Paragraph/NSMutableParagraphStyle+PickerValues.swift
@@ -1,5 +1,5 @@
 //
-//  NSMutableParagraphStyle+Picker.swift
+//  NSMutableParagraphStyle+PickerValues.swift
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2025-04-04.