+++ Sources/RichTextKit/Paragraph/NSParagraphStyle+Views.swift
@@ -1,5 +1,5 @@
 //
-//  NSParagraphStyle+Presentation.swift
+//  NSParagraphStyle+Views.swift
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2025-04-04.